Business Tracking Application
It is a vindows form project with user input according to the authorization Id. There are two Authorization Names in Admin and Staff Names. Admin has the authority to view all pages and take action on all pages. On the other hand, personnel can be viewed as limited to their own areas, and the work they will do is limited. Tcp communication is available. Necessary database is available in files